requirements.txt
fastapi==0.115.2
uvicorn==0.30.6
aiogram==3.12.0
aiosqlite==0.19.0
sqlalchemy==2.0.36
pydantic==2.8.2
pytest==8.2.1
httpx==0.27.2
python-dotenv==1.0.1
jinja2==3.1.4
python-multipart==0.0.9
# Offline ASR and audio conversion
vosk==0.3.45
imageio-ffmpeg==0.4.9